类型转换

+++ CProcessStatus.h
@@ -64,7 +64,7 @@ private:
 
 	int m_iIntval;
 
-	int m_iFileMaxCount;
+	size_t m_iFileMaxCount;
 public:
 
 	int Init(const std::string & dir,int intval=3000,const std::string & hra_file_name = ".hra.processing.file",int max_file_count=100)
